Move CookieConsent out of laikit
