+++ src/components/CookieConsent/index.tsx
